Add support for allocation of old generation in storage device
This commit support the allocation of Old generation of Parallel
Scavenge Heap (ParallelGC) on a raw block storage device using mmap or
fmap. This support works only for ParallelGC. To use this operation use
the following flags:
    - XX:AllocateOldGenAt=</path/to/allocate/old/generation>
    - Xmn<size of new generation>
    - Xmx<total heap size>

Remember to uncomment lines on os_posix.cpp to support allocation on
storage device using filesystem.