[fix] Resolve performance bug in C2 compiler
Problem:
The code we added in C2 to check if the object belongs to TeraCache or
in the Java heap to mark the appropriate card table was wrong, resulting
in  C2 compilation fails. Then the interpreter compiled the method
(slow). The problem occured by the two stores and sync calls that we
added.

Solution:
We resolve the bug by using only one store and one sync call.