[fix] Resolve bug in C2 compiler
Problem: The code we added in C2 to check if the object belongs to
TeraCache or in the Java heap and mark the appropriate card table did
not generate the assembly code for the two paths. Also the comparison
between unsigned ptrs was wrong.

Solution:
Now we support two version of code in C2 when TeraCache is enabled:
    1. Generate inline assembly for the heap card table and make a leaf
    call to mark tera card table.
    2. Make only one leaf call and inside the called function mark the
    appropriate card table. To enable this version of code just enable
    the 'C2_ONLY_LEAF_CALL' in sharedDefines.h