[feat] Add support for fast map hybrid version
Fastmap supports a hybrid version with regular and huge pages. For this
reason we give hints to the kernel using madvise when to use regular and
when to use huge pages. At this time for mutator threads we use huge
pages and for garbage collector threads we use regular pages.

Enable this feature by setting FMAP_HYBRID flag in sharedDefines.h