[feat] Extend object allocation policy for TeraCache
We extend the Teraflag word to locate both RDD Id and the object
state during marking process by the Spark framework. We locate the RDD
Id to the upper 32-bits and the state of the object to the
lower 32-bits of the Tera-flag word. During precompaction phase we
should pass the rdd Id to the TeraCache allocator to give us space in
the appropriate region that locates all the objects of the current rdd
with the specified Id.