[fix][g1full] Fix humongous marking and forwarding
For vanilla G1, Humongous objects were not supposed to move. So,
if they had a value stored in their markWord, the collector would
not preserve it and make it NULL resulting humongous object
showing as they are forwarded and thus not take address from
teraheap. So we explicitly preserve them and check if they
have and H2 address instead of just seeing if they are forwarded.